# 02-Minimal-APIs/JusticeLeague/JusticeLeague.Api/wwwroot/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JusticeLeague.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:7140/",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/heroes/{id}": {
      "get": {
        "tags": [
          "Heroes"
        ],
        "operationId": "GetHeroById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hero"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hero"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Hero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "heroName": {
            "type": "string",
            "nullable": true
          },
          "civillianName": {
            "type": "string",
            "nullable": true
          },
          "superpower": {
            "$ref": "#/components/schemas/Superpower"
          }
        },
        "additionalProperties": false
      },
...
      "Superpower": {
        "enum": [
          "SuperStrength",
          "Flight",
          "SuperSpeed"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JSON Web Token based security",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}